Create README.md
I jhhad codiv 19 and spent 4 months in a nursing home. While there I had a heart attack and pneumonia.I have built websites but forgot what I knew after learning to walk. talk, feed myself and  now in a wheelchair. I am trying to finish a page with a form on it but am stuck.I have the form opening the Email but does not transfer the results to the Email. It arrives with no body.